Upgraded the framework version to 1.3.1

+++ update.sh
@@ -6,77 +6,52 @@
 
 #Update KitHack
 
+# colors
+red='\e[1;31m'
+default='\e[0m'
+yellow='\e[0;33m'
+orange='\e[38;5;166m'
+green='\033[92m'
+
+# location
 path=$(pwd)
 
 # Check root 
 if [ "$(id -u)" != "0" ] > /dev/null 2>&1; then
-echo ""
-echo -e '\e[1;31m[x] Este script necesita permisos root.\e[0m' 1>&2
-sleep 2
+echo -e '\n$red[x] Este script necesita permisos root.' 1>&2
 exit
 fi
 
 # Banner
 clear
 sleep 2
-echo -e "\e[0;33m ____ ___            .___                                 "
-echo -e "\e[0;33m|    |   \______   __| _/____ _/  |_  ____                "
-echo -e "\e[0;33m|    |   /\____ \ / __ |\__  \\   __\/ __ \               "
-echo -e "\e[0;33m|    |  / |  |_> > /_/ | / __ \|  | \  ___/               "
-echo -e "\e[0;33m|______/  |   __/\____ |(____  /__|  \___  > /\  /\  /\   "
-echo -e "\e[0;33m          |__|        \/     \/          \/  \/  \/  \/   "
-echo -e ""
-echo -e "                \e[38;5;166m Update KitHack v1.3.0                "
-echo -e ""
-echo -e "                      By:AdrMXR                                   "
+echo -e "$yellow  ____ ___            .___                                 "
+echo -e "$yellow |    |   \______   __| _/____ _/  |_  ____                "
+echo -e "$yellow |    |   /\____ \ / __ |\__  \\   __\/ __ \               "
+echo -e "$yellow |    |  / |  |_> > /_/ | / __ \|  | \  ___/               "
+echo -e "$yellow |______/  |   __/\____ |(____  /__|  \___  > /\  /\  /\   "
+echo -e "$yellow           |__|        \/     \/          \/  \/  \/  \/   "
+echo -e "                                                                  "
+echo -e "$orange                    Update KitHack v1.3.1                  "
+echo -e "                                                                  "
+echo -e "$orange                         By:AdrMXR                         "
 
 # Check if there is an internet connection
 ping -c 1 google.com > /dev/null 2>&1
 if [[ "$?" == 0 ]]; then
 echo ""
-echo -e "\033[92m[✔][Internet Connection]............[ OK ]"
+echo -e "$green[✔][Internet Connection]............[ OK ]"
 sleep 1.5
 else
 echo ""
-echo -e "\e[1;31m[!][Internet Connection].........[ NOT FOUND ]"
+echo -e "$red[!][Internet Connection].........[ NOT FOUND ]"
 echo ""
 exit
 fi
 
-# Check if git exists
-echo -e "\e[0;33m"
-echo -n [*] Checando git...= ;
-sleep 3 & while [ "$(ps a | awk '{print $1}' | grep $!)" ] ; do for X in '-' '\' '|' '/'; do echo -en "\b$X"; sleep 0.1; done; done 
-echo ""
-which git > /dev/null 2>&1
-if [ "$?" -eq "0" ]; then
-echo ""
-echo -e "\033[92m[✔][GIT]............................[ OK ]"
-sleep 1.5
-else
-echo -e "\e[1;31m[!][GIT].........................[ NOT FOUND ]"
-sleep 1.5
-echo -e "\e[0;33m[!][Installing GIT...]"
-xterm -T "INSTALLER GIT" -geometry 100x30 -e "apt-get install git -y"
-fi 
-
-# Checking python requirements
-echo -e "\e[0;33m"
-echo -n [*] Verificando requerimientos de python...= ;
-sleep 3 & while [ "$(ps a | awk '{print $1}' | grep $!)" ] ; do for X in '-' '\' '|' '/'; do echo -en "\b$X"; sleep 0.1; done; done
-echo ""
-echo -e "\033[92m"
-pip3 install requests
-pip3 install py-getch
-apt-get install python3-tk
-pip3 install pathlib
-pip3 install zenipy
-pip3 install pgrep
-apt-get install libatk-adaptor libgail-common
-sudo apt-get purge fcitx-module-dbus
-
 # Creating temporary directory...
-echo -e "\e[0;33m"
+if [ -d tools ] && [ -d output ]; then
+echo -e "$yellow"
 echo -n [*] Creando directorio temporal...= ;
 sleep 3 & while [ "$(ps a | awk '{print $1}' | grep $!)" ] ; do for X in '-' '\' '|' '/'; do echo -en "\b$X"; sleep 0.1; done; done 
 echo ""
@@ -85,29 +60,38 @@ mkdir "$temp"
 mv "$path/output" "$temp/output" 
 mv "$path/tools" "$temp/tools"
 echo ""
-echo -e "\033[92m[✔] Done."
+echo -e "$green[✔] Done."
 sleep 1.5
+fi
 
 # Updating KitHack...
-echo -e "\e[0;33m"
-echo -n [*] Actualizando repositorio KitHack desde Github...= ;
+echo -e "$yellow"
+echo -n [*] Actualizando KitHack desde Github...= ;
 sleep 3 & while [ "$(ps a | awk '{print $1}' | grep $!)" ] ; do for X in '-' '\' '|' '/'; do echo -en "\b$X"; sleep 0.1; done; done 
-echo -e "\033[92m"
+echo -e "$green"
 echo ""
+git config --global pull.rebase true
 git reset HEAD --hard
-git pull
-echo ""
+git pull origin master
 sleep 1.5
 
 # Moving the files in the temporary directory again
-echo -e "\e[0;33m"
-echo -n [*] Moviendo de nuevo los archivos del directorio temporal...= ;
+if [ -d $HOME/hacktemp ]; then
+echo -e "$yellow"
+echo -n [*] Moviendo archivos del directorio temporal...= ;
 sleep 3 & while [ "$(ps a | awk '{print $1}' | grep $!)" ] ; do for X in '-' '\' '|' '/'; do echo -en "\b$X"; sleep 0.1; done; done 
 echo ""
 mv "$temp/output" "$path/output"
 mv "$temp/tools" "$path/tools"
 rm -rf "$temp"
 echo ""
-echo -e "\033[92m[✔] Done. Update complete."
+echo -e "$green[✔] Done."
 sleep 1.5
-exit 0
+fi
+
+# Running installer
+echo -e "$yellow"
+echo -n [*] Ejecutando instalador...= ;
+sleep 3 & while [ "$(ps a | awk '{print $1}' | grep $!)" ] ; do for X in '-' '\' '|' '/'; do echo -en "\b$X"; sleep 0.1; done; done 
+bash install.sh
+exit
